+++ pkgs/dart_mcp/lib/src/server/server.dart
@@ -4,9 +4,8 @@
 
 import 'dart:async';
 
-import 'package:json_rpc_2/json_rpc_2.dart';
 import 'package:meta/meta.dart';
-import 'package:stream_channel/stream_channel.dart';
+import 'package:stream_transform/stream_transform.dart';
 
 import '../api/api.dart';
 import '../shared.dart';
@@ -15,6 +14,7 @@ part 'completions_support.dart';
 part 'logging_support.dart';
 part 'prompts_support.dart';
 part 'resources_support.dart';
+part 'roots_tracking_support.dart';
 part 'tools_support.dart';
 
 /// Base class to extend when implementing an MCP server.
@@ -59,11 +59,12 @@ abstract base class MCPServer extends MCPBase {
       _rootsListChangedController?.stream;
   StreamController<RootsListChangedNotification>? _rootsListChangedController;
 
-  MCPServer.fromStreamChannel({
+  MCPServer.fromStreamChannel(
+    super.channel, {
     required this.implementation,
     required this.instructions,
-    required StreamChannel<String> channel,
-  }) : super(Peer(channel)) {
+    super.protocolLogSink,
+  }) {
     registerRequestHandler(InitializeRequest.methodName, initialize);
 
     registerNotificationHandler(