+++ pkgs/dart_tooling_mcp_server/lib/src/mixins/analyzer.dart
@@ -13,12 +13,14 @@ import 'package:language_server_protocol/protocol_generated.dart' as lsp;
 import 'package:meta/meta.dart';
 
 import '../lsp/wire_format.dart';
+import '../utils/constants.dart';
 
 /// Mix this in to any MCPServer to add support for analyzing Dart projects.
 ///
 /// The MCPServer must already have the [ToolsSupport] and [LoggingSupport]
 /// mixins applied.
-base mixin DartAnalyzerSupport on ToolsSupport, LoggingSupport {
+base mixin DartAnalyzerSupport
+    on ToolsSupport, LoggingSupport, RootsTrackingSupport {
   /// The LSP server connection for the analysis server.
   late final Peer _lspConnection;
 
@@ -32,30 +34,37 @@ base mixin DartAnalyzerSupport on ToolsSupport, LoggingSupport {
   /// is over.
   Completer<void>? _doneAnalyzing = Completer();
 
-  /// All known workspace roots.
-  ///
-  /// Identity is controlled by the [Root.uri].
-  Set<Root> workspaceRoots = HashSet(
-    equals: (r1, r2) => r2.uri == r2.uri,
-    hashCode: (r) => r.uri.hashCode,
-  );
+  /// The current LSP workspace folder state.
+  HashSet<lsp.WorkspaceFolder> _currentWorkspaceFolders =
+      HashSet<lsp.WorkspaceFolder>(
+        equals: (a, b) => a.uri == b.uri,
+        hashCode: (a) => a.uri.hashCode,
+      );
 
   @override
   FutureOr<InitializeResult> initialize(InitializeRequest request) async {
+    // This should come first, assigns `clientCapabilities`.
+    final result = await super.initialize(request);
+
     // We check for requirements and store a message to log after initialization
     // if some requirement isn't satisfied.
-    var unsupportedReason =
-        request.capabilities.roots == null
-            ? 'Project analysis requires the "roots" capability which is not '
-                'supported. Analysis tools have been disabled.'
-            : (Platform.environment['DART_SDK'] == null
-                ? 'Project analysis requires a "DART_SDK" environment variable '
-                    'to be set (this should be the path to the root of the '
-                    'dart SDK). Analysis tools have been disabled.'
-                : null);
+    final unsupportedReasons = <String>[
+      if (!supportsRoots)
+        'Project analysis requires the "roots" capability which is not '
+            'supported. Analysis tools have been disabled.',
+      if (Platform.environment['DART_SDK'] == null)
+        'Project analysis requires a "DART_SDK" environment variable to be set '
+            '(this should be the path to the root of the dart SDK). Analysis '
+            'tools have been disabled.',
+    ];
 
-    unsupportedReason ??= await _initializeAnalyzerLspServer();
-    if (unsupportedReason == null) {
+    if (unsupportedReasons.isEmpty) {
+      if (await _initializeAnalyzerLspServer() case final failedReason?) {
+        unsupportedReasons.add(failedReason);
+      }
+    }
+
+    if (unsupportedReasons.isEmpty) {
       registerTool(analyzeFilesTool, _analyzeFiles);
       registerTool(resolveWorkspaceSymbolTool, _resolveWorkspaceSymbol);
     }
@@ -64,16 +73,13 @@ base mixin DartAnalyzerSupport on ToolsSupport, LoggingSupport {
     // (even logging).
     unawaited(
       initialized.then((_) {
-        if (unsupportedReason != null) {
-          log(LoggingLevel.warning, unsupportedReason);
-        } else {
-          // All requirements satisfied, ask the client for its roots.
-          _listenForRoots();
+        if (unsupportedReasons.isNotEmpty) {
+          log(LoggingLevel.warning, unsupportedReasons.join('\n'));
         }
       }),
     );
 
-    return super.initialize(request);
+    return result;
   }
 
   /// Initializes the analyzer lsp server.
@@ -233,12 +239,14 @@ base mixin DartAnalyzerSupport on ToolsSupport, LoggingSupport {
   ///
   /// Waits for any pending analysis before returning.
   Future<CallToolResult> _analyzeFiles(CallToolRequest request) async {
-    await _doneAnalyzing?.future;
+    final errorResult = await _ensurePrerequisites(request);
+    if (errorResult != null) return errorResult;
+
     final messages = <Content>[];
     for (var entry in diagnostics.entries) {
       for (var diagnostic in entry.value) {
         final diagnosticJson = diagnostic.toJson();
-        diagnosticJson['uri'] = entry.key.toString();
+        diagnosticJson[ParameterNames.uri] = entry.key.toString();
         messages.add(TextContent(text: jsonEncode(diagnosticJson)));
       }
     }
@@ -253,13 +261,28 @@ base mixin DartAnalyzerSupport on ToolsSupport, LoggingSupport {
   Future<CallToolResult> _resolveWorkspaceSymbol(
     CallToolRequest request,
   ) async {
-    await _doneAnalyzing?.future;
-    final query = request.arguments!['query'] as String;
+    final errorResult = await _ensurePrerequisites(request);
+    if (errorResult != null) return errorResult;
+
+    final query = request.arguments![ParameterNames.query] as String;
     final result = await _lspConnection.sendRequest(
       lsp.Method.workspace_symbol.toString(),
       lsp.WorkspaceSymbolParams(query: query).toJson(),
     );
     return CallToolResult(content: [TextContent(text: jsonEncode(result))]);
+  }
+
+  /// Ensures that all prerequisites for any analysis task are met.
+  ///
+  /// Returns an error response if any prerequisite is not met, otherwise
+  /// returns `null`.
+  Future<CallToolResult?> _ensurePrerequisites(CallToolRequest request) async {
+    final roots = await this.roots;
+    if (roots.isEmpty) {
+      return noRootsSetResponse;
+    }
+    await _doneAnalyzing?.future;
+    return null;
   }
 
   /// Handles `$/analyzerStatus` events, which tell us when analysis starts and
@@ -284,47 +307,53 @@ base mixin DartAnalyzerSupport on ToolsSupport, LoggingSupport {
     );
     diagnostics[diagnosticParams.uri] = diagnosticParams.diagnostics;
     log(LoggingLevel.debug, {
-      'uri': diagnosticParams.uri,
+      ParameterNames.uri: diagnosticParams.uri,
       'diagnostics':
           diagnosticParams.diagnostics.map((d) => d.toJson()).toList(),
     });
   }
 
-  /// Lists the roots, and listens for changes to them.
-  ///
-  /// Sends workspace change notifications to the LSP server based on the roots.
-  void _listenForRoots() async {
-    rootsListChanged!.listen((event) async {
-      await _updateRoots();
-    });
-    await _updateRoots();
-  }
+  /// Update the LSP workspace dirs when our workspace [Root]s change.
+  @override
+  Future<void> updateRoots() async {
+    await super.updateRoots();
+    unawaited(() async {
+      final newRoots = await roots;
 
-  /// Updates the set of [workspaceRoots] and notifies the server.
-  Future<void> _updateRoots() async {
-    final newRoots = HashSet<Root>(
-      equals: (r1, r2) => r1.uri == r2.uri,
-      hashCode: (r) => r.uri.hashCode,
-    )..addAll((await listRoots(ListRootsRequest())).roots);
+      final oldWorkspaceFolders = _currentWorkspaceFolders;
+      final newWorkspaceFolders =
+          _currentWorkspaceFolders = HashSet<lsp.WorkspaceFolder>(
+            equals: (a, b) => a.uri == b.uri,
+            hashCode: (a) => a.uri.hashCode,
+          )..addAll(newRoots.map((r) => r.asWorkspaceFolder));
 
-    final removed = workspaceRoots.difference(newRoots);
-    final added = newRoots.difference(workspaceRoots);
-    workspaceRoots = newRoots;
+      final added =
+          newWorkspaceFolders.difference(oldWorkspaceFolders).toList();
+      final removed =
+          oldWorkspaceFolders.difference(newWorkspaceFolders).toList();
 
-    final event = lsp.WorkspaceFoldersChangeEvent(
-      added: [for (var root in added) root.asWorkspaceFolder],
-      removed: [for (var root in removed) root.asWorkspaceFolder],
-    );
+      // This can happen in the case of multiple notifications in quick
+      // succession, the `roots` future will complete only after the state has
+      // stabilized which can result in empty diffs.
+      if (added.isEmpty && removed.isEmpty) {
+        return;
+      }
 
-    log(
-      LoggingLevel.debug,
-      () => 'Notifying of workspace root change: ${event.toJson()}',
-    );
+      final event = lsp.WorkspaceFoldersChangeEvent(
+        added: added,
+        removed: removed,
+      );
 
-    _lspConnection.sendNotification(
-      lsp.Method.workspace_didChangeWorkspaceFolders.toString(),
-      lsp.DidChangeWorkspaceFoldersParams(event: event).toJson(),
-    );
+      log(
+        LoggingLevel.debug,
+        () => 'Notifying of workspace root change: ${event.toJson()}',
+      );
+
+      _lspConnection.sendNotification(
+        lsp.Method.workspace_didChangeWorkspaceFolders.toString(),
+        lsp.DidChangeWorkspaceFoldersParams(event: event).toJson(),
+      );
+    }());
   }
 
   @visibleForTesting
@@ -332,6 +361,7 @@ base mixin DartAnalyzerSupport on ToolsSupport, LoggingSupport {
     name: 'analyze_files',
     description: 'Analyzes the entire project for errors.',
     inputSchema: Schema.object(),
+    annotations: ToolAnnotations(title: 'Analyze projects', readOnlyHint: true),
   );
 
   @visibleForTesting
@@ -340,16 +370,28 @@ base mixin DartAnalyzerSupport on ToolsSupport, LoggingSupport {
     description: 'Look up a symbol or symbols in all workspaces by name.',
     inputSchema: Schema.object(
       properties: {
-        'query': Schema.string(
+        ParameterNames.query: Schema.string(
           description:
               'Queries are matched based on a case-insensitive partial name '
               'match, and do not support complex pattern matching, regexes, '
               'or scoped lookups.',
         ),
       },
-      required: ['query'],
+      required: [ParameterNames.query],
     ),
     annotations: ToolAnnotations(title: 'Project search', readOnlyHint: true),
+  );
+
+  @visibleForTesting
+  static final noRootsSetResponse = CallToolResult(
+    isError: true,
+    content: [
+      TextContent(
+        text:
+            'No roots set. At least one root must be set in order to use this '
+            'tool.',
+      ),
+    ],
   );
 }