+++ pkgs/dart_tooling_mcp_server/lib/src/mixins/dart_cli.dart
@@ -7,6 +7,7 @@ import 'dart:async';
 import 'package:dart_mcp/server.dart';
 
 import '../utils/cli_utils.dart';
+import '../utils/constants.dart';
 import '../utils/process_manager.dart';
 
 // TODO: migrate the analyze files tool to use this mixin and run the
@@ -17,18 +18,19 @@ import '../utils/process_manager.dart';
 ///
 /// The MCPServer must already have the [ToolsSupport] and [LoggingSupport]
 /// mixins applied.
-base mixin DartCliSupport on ToolsSupport, LoggingSupport
+base mixin DartCliSupport on ToolsSupport, LoggingSupport, RootsTrackingSupport
     implements ProcessManagerSupport {
   @override
   FutureOr<InitializeResult> initialize(InitializeRequest request) {
-    if (request.capabilities.roots == null) {
-      throw StateError(
-        'This server requires the "roots" capability to be implemented.',
-      );
+    try {
+      return super.initialize(request);
+    } finally {
+      // Can't call this until after `super.initialize`.
+      if (supportsRoots) {
+        registerTool(dartFixTool, _runDartFixTool);
+        registerTool(dartFormatTool, _runDartFormatTool);
+      }
     }
-    registerTool(dartFixTool, _runDartFixTool);
-    registerTool(dartFormatTool, _runDartFormatTool);
-    return super.initialize(request);
   }
 
   /// Implementation of the [dartFixTool].
@@ -38,6 +40,7 @@ base mixin DartCliSupport on ToolsSupport, LoggingSupport
       command: ['dart', 'fix', '--apply'],
       commandDescription: 'dart fix',
       processManager: processManager,
+      knownRoots: await roots,
     );
   }
 
@@ -49,59 +52,25 @@ base mixin DartCliSupport on ToolsSupport, LoggingSupport
       commandDescription: 'dart format',
       processManager: processManager,
       defaultPaths: ['.'],
+      knownRoots: await roots,
     );
   }
 
   static final dartFixTool = Tool(
     name: 'dart_fix',
     description: 'Runs `dart fix --apply` for the given project roots.',
-    inputSchema: ObjectSchema(
-      properties: {
-        'roots': ListSchema(
-          title: 'All projects roots to run dart fix in.',
-          description:
-              'These must match a root returned by a call to "listRoots".',
-          items: ObjectSchema(
-            properties: {
-              'root': StringSchema(
-                title: 'The URI of the project root to run `dart fix` in.',
-              ),
-            },
-            required: ['root'],
-          ),
-        ),
-      },
-      required: ['roots'],
+    annotations: ToolAnnotations(title: 'Dart fix', destructiveHint: true),
+    inputSchema: Schema.object(
+      properties: {ParameterNames.roots: rootsSchema()},
     ),
   );
 
   static final dartFormatTool = Tool(
     name: 'dart_format',
     description: 'Runs `dart format .` for the given project roots.',
-    inputSchema: ObjectSchema(
-      properties: {
-        'roots': ListSchema(
-          title: 'All projects roots to run dart format in.',
-          description:
-              'These must match a root returned by a call to "listRoots".',
-          items: ObjectSchema(
-            properties: {
-              'root': StringSchema(
-                title: 'The URI of the project root to run `dart format` in.',
-              ),
-              'paths': ListSchema(
-                title:
-                    'Relative or absolute paths to analyze under the '
-                    '"root". Paths must correspond to files and not '
-                    'directories.',
-                items: StringSchema(),
-              ),
-            },
-            required: ['root'],
-          ),
-        ),
-      },
-      required: ['roots'],
+    annotations: ToolAnnotations(title: 'Dart format', destructiveHint: true),
+    inputSchema: Schema.object(
+      properties: {ParameterNames.roots: rootsSchema(supportsPaths: true)},
     ),
   );
 }